requirements_v3.txt
# === CPU models (轻量感知层) ===
deepface>=0.0.92          # 摄像头表情检测 (~50MB model)
tf-keras                  # DeepFace 依赖
transformers>=4.30.0      # 文字情感分析 DistilBERT
torch                     # transformers 依赖（CPU版即可）
sentencepiece             # tokenizer 依赖

# === 音频 (邱靖翔负责) ===
funasr                   # emotion2vec 语音情绪
modelscope               # emotion2vec 模型下载/管理
panns-inference          # 环境音分类
torchaudio               # PANNs 音频加载依赖
openai-whisper            # 语音转文字

# === TTS (陈昕博负责) ===
# ChatTTS                 # 本地语音合成（替代 edge-tts）
edge-tts                  # 临时使用，最终换成 ChatTTS

# === 基础 ===
opencv-python-headless>=4.8.0
numpy>=1.24.0
requests>=2.31.0
flask>=3.0.0

# === RAG (邱靖翔负责) ===
chromadb>=0.4.0
sentence-transformers>=2.2.0
